Refactor ws2812_stream module for improved state management and bit counting
- Introduced state machine with states: ST_RESET, ST_LOAD, and ST_SEND.
- Updated reset cycle count to 7679 for 60 us duration at 128 MHz.
- Changed bit cell count from 6 bits to 8 bits for better resolution.
- Adjusted high count calculation for data output based on new bit cell count.
- Streamlined the logic for loading and sending data to the WS2812 LED strip.

+++ Core/Inc/device_profile.h
@@ -32,7 +32,7 @@ typedef struct __attribute__((packed)) {
 } umh_element_coordinate_t;
 
 typedef struct __attribute__((packed)) {
-  uint16_t phase;
+  uint8_t phase;
   uint8_t gain;
   uint8_t enabled;
 } umh_channel_calibration_t;
